WAC 74 - [TMP] Added a view file for extra_vars
